feat: integrate ICodecBackend into render pipeline for frame decoding
- Update `RenderGraph` and `IRenderNode` to accept `ICodecBackend` during execution.
- Implement frame fetching in `SequenceNode` using calculated source time and the codec backend.
- Update `SequenceNode::setClips` to use `ResolvedClip` for mapping timeline positions to source media.
- Fix `Time` deserialization to read `timebase` as an `int64_t`.
- Enhance `MockGpuBackend` to support `clear_black` and `lut_3d` compute dispatch in tests.
- Add `MockCodecBackend` to internal tests to support the updated render signature.

+++ core/src/ffi/project_ffi.cpp
@@ -92,7 +92,7 @@ vx_status_t vx_project_add_asset(VxProject* project, const char* path, int64_t d
         return VX_ERR_INVALID_ARG;
     }
 
-    vx::Time duration = vx::Time::fromSeconds(duration_ms / 1000.0);
+    vx::Time duration = vx::Time::fromSeconds(static_cast<double>(duration_ms) / 1000.0);
     vx::Resolution resolution{width, height};
     projectService().addMediaAsset(*project->project, path, duration, resolution);
     return VX_OK;
